Update README.md
The program is designed to open and read an input file which contains an unknown number of records. It is then to print the list of IDs and test scores. Once the info is printed, the program will sort the lists by ID number low to high and print it again. Then it will calculate the average test score and determine the overall grade of the score and print again. The program will then sort the list based on test scores from highest to lowest and print a final list.